Grid scaling alignment working, still need to fill in grid more when scaled down

+++ editor_tree_inspector_update.cpp
@@ -63,7 +63,7 @@ void TreeInspector::updateInspectorPropertyBoxes(QList<DrSettings*> changed_item
         case Property_Type::Int:
         case Property_Type::Positive:   dynamic_cast<QSpinBox*>(widget)->setValue(prop->getValue().toInt());            break;
 
-        case Property_Type::Float:
+        case Property_Type::Double:
         case Property_Type::Percent:
         case Property_Type::Angle:      dynamic_cast<QDoubleSpinBox*>(widget)->setValue(prop->getValue().toDouble());   break;
 
@@ -136,7 +136,7 @@ void TreeInspector::updateSettingsFromNewValue(long property_key, QVariant new_v
         case Property_Type::Bool:                                   // true or false
         case Property_Type::Int:                                    // any integer
         case Property_Type::Positive:                               // integer >= 0
-        case Property_Type::Float:                                  // any floating point
+        case Property_Type::Double:                                 // any floating point
         case Property_Type::Percent:                                // floating point from 0.0 to 100.0
         case Property_Type::Angle:                                  // floating point for showing degrees
         case Property_Type::List:                                   // index value
