+++ project_world.cpp
@@ -89,9 +89,9 @@ void DrWorld::initializeWorldSettings(QString new_name)
     getComponent(Components::World_Settings)->setIcon(Component_Icons::Settings);
     addPropertyToComponent(Components::World_Settings, Properties::World_Name, Property_Type::String, new_name,
                            "World Name", "Name of the current world.");
-    addPropertyToComponent(Components::World_Settings, Properties::World_Game_Direction, Property_Type::Float, 0.0,
+    addPropertyToComponent(Components::World_Settings, Properties::World_Game_Direction, Property_Type::Double, 0.0,
                            "Game Direction", "Default direction, in degrees, to load new stages, 0 - right, 90 - up, 180 - left, 270 - down, etc.");
-    addPropertyToComponent(Components::World_Settings, Properties::World_Score_Multiplier, Property_Type::Float, 1.0,
+    addPropertyToComponent(Components::World_Settings, Properties::World_Score_Multiplier, Property_Type::Double, 1.0,
                            "Score Multiplier", "Value used as multiplier to adjust speed at which distance scoring is calculated.");
 
     addComponent(Components::World_Physics, "Physics", "Starting physics settings for current world.", Component_Colors::Orange_Medium, true);
@@ -100,13 +100,13 @@ void DrWorld::initializeWorldSettings(QString new_name)
                            "Use Physics?", "Whether or not physics is turned on in current world.");
     addPropertyToComponent(Components::World_Physics, Properties::World_Gravity, Property_Type::PointF, QPointF(0.0, 10.0),
                            "Gravity", "Amount of gravity in x and y directions, can be negative.");
-    addPropertyToComponent(Components::World_Physics, Properties::World_Time_Warp, Property_Type::Float, 60.0,
+    addPropertyToComponent(Components::World_Physics, Properties::World_Time_Warp, Property_Type::Double, 60.0,
                            "Time Warp", "Update calculations per second.");
-    addPropertyToComponent(Components::World_Physics, Properties::World_Friction, Property_Type::Float, 10.0,
+    addPropertyToComponent(Components::World_Physics, Properties::World_Friction, Property_Type::Double, 10.0,
                            "Friction", "Global friction setting.");
-    addPropertyToComponent(Components::World_Physics, Properties::World_Drag, Property_Type::Float, 5.0,
+    addPropertyToComponent(Components::World_Physics, Properties::World_Drag, Property_Type::Double, 5.0,
                            "Velocity Drag", "Velocity drag.");
-    addPropertyToComponent(Components::World_Physics, Properties::World_Bounce, Property_Type::Float, 5.0,
+    addPropertyToComponent(Components::World_Physics, Properties::World_Bounce, Property_Type::Double, 5.0,
                            "Bounce", "Global bounce setting.");
 
 }
